Agregado de toda la configuracion de la pantalla de clientes

+++ app/Http/Controllers/ClienteController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Cliente;
+use Illuminate\Support\Facades\DB;
 
 class ClienteController extends Controller
 {
@@ -12,10 +13,27 @@ class ClienteController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $clientes = Cliente::all();
-        return $clientes;
+        if(!$request->ajax()) return redirect('/');
+        $buscar = $request->buscar;
+        $criterio = $request->criterio;
+        if($buscar == ''){
+            $clientes =  Cliente::orderBy('id','desc')->paginate(5);
+        }else{
+            $clientes =  Cliente::where($criterio,'like','%'.$buscar.'%')->orderBy('id','desc')->paginate(5);
+        }
+        return [
+            'pagination'=>[
+                'total'=>$clientes->total(),
+                'current_page'=>$clientes->currentPage(),
+                'per_page'=>$clientes->perPage(),
+                'last_page'=>$clientes->lastPage(),
+                'from'=>$clientes->firstItem(),
+                'to'=>$clientes->lastItem(),
+            ],
+            'clientes'=>$clientes
+        ];
     }
 
     /**
@@ -37,15 +55,18 @@ class ClienteController extends Controller
     public function store(Request $request)
     {
         //
+        if(!$request->ajax()) return redirect('/');
         $cliente = new Cliente();
         $cliente->zona = $request->zona;
         $cliente->nombre = $request->nombre;
         $cliente->avenida = $request->avenida;
+        $cliente->calle = $request->calle;
+        $cliente->pisos = $request->pisos;
         $cliente->distancia = $request->distancia;
         $cliente->cantidad = $request->cantidad;
-        $cliente->latitud = $request->longitud;
+        $cliente->latitud = $request->latitud;
+        $cliente->longitud = $request->longitud;
         $cliente->razonSocial = $request->razonSocial;
-        $cliente->nombre = $request->nombre;
         $cliente->telefono = $request->telefono;
         $cliente->observacion = $request->observacion;
         $cliente->save();
@@ -60,7 +81,6 @@ class ClienteController extends Controller
      */
     public function show($id)
     {
-        //
     }
 
     /**
@@ -71,7 +91,6 @@ class ClienteController extends Controller
      */
     public function edit($id)
     {
-        //
     }
 
     /**
@@ -83,10 +102,12 @@ class ClienteController extends Controller
      */
     public function update(Request $request)
     {
-        $cliente = new Cliente::findOrFail($request->id);
+        if(!$request->ajax()) return redirect('/');
+        $cliente = Cliente::findOrFail($request->id);
         $cliente->zona = $request->zona;
         $cliente->nombre = $request->nombre;
         $cliente->avenida = $request->avenida;
+        $cliente->pisos = $request->pisos;
         $cliente->distancia = $request->distancia;
         $cliente->cantidad = $request->cantidad;
         $cliente->latitud = $request->longitud;
@@ -103,8 +124,9 @@ class ClienteController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($request)
+    public function destroy(Request $request)
     {
+        if(!$request->ajax()) return redirect('/');
         Cliente::destroy($request->id);
     }
 }
